elif ,if
elif in Python stands for "else if" and is used for conditional statements. It allows you to check multiple expressions for truth and execute a block of code as soon as one of the conditions is met. If the if condition is false, the program will check the elif condition(s), and if none of the elif conditions are true, it can execute an optional else block.